+++ changelog.md
@@ -11,6 +11,10 @@
 ### [Mods Base v1.10](https://github.com/bl-sdk/mods_base/blob/master/Readme.md#v19)
 > - Linting fixups.
 
+### Save Options v1.2
+- Fixed issue where loading a character with no save option data inherited option values from
+  previous character.
+
 ### UI Utils v1.3
 - Linting fixes.